+++ walkers/src/lib.rs
@@ -15,5 +15,5 @@ mod zoom;
 pub use download::{HeaderValue, HttpOptions};
 pub use map::{Map, MapMemory, Plugin, Projector};
 pub use mercator::{screen_to_position, Position, TileId};
-pub use tiles::{Texture, Tiles, TilesManager};
+pub use tiles::{HttpTiles, Texture, Tiles};
 pub use zoom::InvalidZoom;